fix: conditional OOB method calls for MODEXP_LEAD and MODEXP_PRICING
We make the OOB method cals to OOB_INST_MODEXP_LEAD and
OOB_INST_MODEXP_PRICING conditional to

    all_byte_sizes_are_in_bounds =
                  bbs_is_in_bounds
                ∧ ebs_is_in_bounds
                ∧ mbs_is_in_bounds

Otherwise we had issues in the tracer implementation getting the correct
exponent lead word (as we were reading call data starting at 96 + bbs in
some parts of the tracer, and 96 + normalizedBbs in others. So this spec
change just makes life easier, and avoids computing exponent logs that
will never be used.